pyproject.toml
[project]
name = "booksapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mariadb>=1.1.13",
    "fastapi>=0.111.0",
    "pydantic>=2.7.4",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.0.8",
    "uvicorn>=0.30.1"
]

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mariadb"
ignore_errors = true
